% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\name{manhattan_plotter}
\alias{manhattan_plotter}
\title{manhattan_plotter}
\usage{
manhattan_plotter(dat, threshold = 1e-07, prefix = "Manhattan plot",
  plotly_prep = FALSE, bi_dir = TRUE, custom_colours = NULL)
}
\arguments{
\item{bi_dir}{}
}
\description{
manhattan_plotter
}
