% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.r
\name{manhattan_pre_proc_analysis}
\alias{manhattan_pre_proc_analysis}
\title{manhattan_pre_proc_analysis
pre-procress and an analysis from an EWAS object in prepatation for a 
manhattan plot}
\usage{
manhattan_pre_proc_analysis(name, ewas, bi_dir = TRUE,
  chromosomes = paste("chr", c(1:22, "X", "Y"), sep = ""))
}
\description{
manhattan_pre_proc_analysis
pre-procress and an analysis from an EWAS object in prepatation for a 
manhattan plot
}
