% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.r
\name{scatter_thinning}
\alias{scatter_thinning}
\title{scatter_thinning}
\usage{
scatter_thinning(x, y, resolution = 100, max_per_cell = 100)
}
\arguments{
\item{x}{x-coordinates vector}

\item{y}{y-coordinates vector}

\item{resolution}{number of partitions for the x and y-dimensions.}

\item{max_per_cell}{maximum number of points per x-y partition.}
}
\value{
index into the points that omits points from x-y partitions
so that each has at most \code{max_per_cell} points.
}
\description{
Remove points from a scatter plot where density is really high
}
